refactor(graph): analyzeSourceFile を独立させる
